Suggestions and Required changes applied
Suggestions and Required changes after Tutors' feedback

SUGGESTION
The word TODO: needs to be removed from the comments as that word in the comments means that something in the code is incomplete. As you have already implemented what was required, reword the comments provided or write your own.

SUGGESTION
Can remove this variable assignment as it is not being used in this test.

SUGGESTION
Try the:   toBeGreaterThan(0)     matcher instead of        >0).toBe(true)


----------------------------------------------------------------------------------------------------
describe('New Feed Selection', function() {

        let feedOne,
            feedTwo;
        
        beforeEach (function(done) {
            loadFeed(0, function(){
                feedOne = $('.feed').html();
                done();
            });
            loadFeed(1,function(){
                feedOne = $('.feed').html();
                done();
            });
        });

        it('Content changes when loadFeed is loaded', function() {
            expect(feedOne === feedTwo).toBe(false);
        });
    });

REQUIRED
Calling done in the first loadFeed() function is telling the program to exit the beforeEach() function and run the spec. So the second loadFeed() function below does not even run.

             loadFeed(1,function(){
                feedOne = $('.feed').html();
                done();

REQUIRED
First, I believe you meant to write feedTwo here.

loadFeed(1,function(){
                feedOne = $('.feed').html();
                done();
            });

Second, this feed is not loading. To ensure that both the feeds load before the spec runs, you must call the second loadFeed() function as a callback of the first.
See here (https://codeburst.io/javascript-what-the-heck-is-a-callback-aba4da2deced) for more information about callbacks.

Example:
load first feed {
    save its contents
     load second feed {
           save its contents
            call done()
       }
}

When loaded like this, the second feed only starts when the first one finishes and the spec runs only after the second one finishes. This way you ensure that both the feeds complete loading before the spec runs.

        it('Content changes when loadFeed is loaded', function() {
            expect(feedOne === feedTwo).toBe(false);

REQUIRED
Right now the test is passing for the wrong reasons. Even if you had correctly used feedTwo instead of feedOne in the second loadFeed function, feedTwo variable will return undefined because the second feed doesn't get a chance to load.. And since feedOne has a value, and since they are not equal, the test is passing. Try fixing the variable name in the second loadFeed() function and logging out the values of the variables as below:

Load the feeds as suggested above to ensure that both variables have the correct value.